personal_trainer_agent/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain-core
langchain-groq
langgraph
